requirements.txt
alembic==0.9.9
click==6.7
Flask==1.0.2
Flask-Cors==3.0.4
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.8.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.4
PyJWT==1.6.1
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.4
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
